# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreatePostRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.Astronaut:
    properties:
      craft:
        type: string
      name:
        type: string
    type: object
  models.AstronautWithPhoto:
    properties:
      imgUrl:
        type: string
      name:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.HistoricalRequest:
    properties:
      timestamp:
        type: integer
      units:
        type: string
    required:
    - timestamp
    type: object
  models.ISSCrewResponse:
    properties:
      message:
        type: string
      people:
        items:
          $ref: '#/definitions/models.Astronaut'
        type: array
    type: object
  models.ISSCrewWithPhotosResponse:
    properties:
      message:
        type: string
      people:
        items:
          $ref: '#/definitions/models.AstronautWithPhoto'
        type: array
    type: object
  models.ISSPosition:
    properties:
      altitude:
        type: number
      created_at:
        type: string
      daynum:
        type: number
      footprint:
        type: number
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      solar_lat:
        type: number
      solar_lon:
        type: number
      timestamp:
        type: integer
      units:
        type: string
      updated_at:
        type: string
      velocity:
        type: number
      visibility:
        type: string
    type: object
  models.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the ISS Model Backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ISS Model Backend API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns a hello world message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hello World
      tags:
      - general
  /admin/blog/posts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Create New Post
      tags:
      - Blog (Admin)
  /admin/blog/posts/{id}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - Blog (Admin)
    put:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Update Post
      tags:
      - Blog (Admin)
  /admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Admin Login
      tags:
      - Auth
  /admin/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register Admin
      tags:
      - Auth
  /blog/posts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: Get All Posts
      tags:
      - Blog
  /blog/posts/{id}:
    get:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      summary: Get Post by ID
      tags:
      - Blog
  /health:
    get:
      consumes:
      - application/json
      description: Returns the health status of the database and application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health Check
      tags:
      - health
  /iss/crew:
    get:
      consumes:
      - application/json
      description: Return current crew aboard the ISS
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ISSCrewResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get current ISS crew
      tags:
      - Crew
  /iss/current:
    get:
      consumes:
      - application/json
      description: Returns the current position of the International Space Station
      parameters:
      - default: kilometers
        description: Units (kilometers or miles)
        enum:
        - kilometers
        - miles
        in: query
        name: units
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ISSPosition'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Current ISS Position
      tags:
      - ISS
  /iss/historical:
    post:
      consumes:
      - application/json
      description: Returns the ISS position for a timestamp provided in request body
      parameters:
      - description: Historical position request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.HistoricalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ISSPosition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Historical ISS Position (POST)
      tags:
      - ISS
  /iss/historical/{timestamp}:
    get:
      consumes:
      - application/json
      description: Returns the ISS position for a specific timestamp (within 4 hours
        back/forward)
      parameters:
      - description: Unix timestamp
        in: path
        name: timestamp
        required: true
        type: integer
      - default: kilometers
        description: Units (kilometers or miles)
        enum:
        - kilometers
        - miles
        in: query
        name: units
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ISSPosition'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Historical ISS Position
      tags:
      - ISS
  /iss/range:
    get:
      consumes:
      - application/json
      description: Returns all ISS positions within a specified time range
      parameters:
      - description: Start timestamp (Unix)
        in: query
        name: start_time
        required: true
        type: integer
      - description: End timestamp (Unix)
        in: query
        name: end_time
        required: true
        type: integer
      - default: kilometers
        description: Units (kilometers or miles)
        enum:
        - kilometers
        - miles
        in: query
        name: units
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ISSPosition'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get ISS Positions in Time Range
      tags:
      - ISS
  /iss/status:
    get:
      consumes:
      - application/json
      description: Returns statistics about ISS tracking data and system status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get ISS Tracking Status
      tags:
      - ISS
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: '"Wpisz ''Bearer '' (ze spacją), a następnie swój token JWT. Przykład:
      ''Bearer eyJhbGciOiJIUzI1Ni...''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"